feat(runtime): migrate all parameter types from runtime to runtime-common
