+++ runtime/common/Cargo.toml
@@ -10,14 +10,26 @@ version = '0.0.1'
 [dependencies]
 frame-support = { git = 'https://github.com/paritytech/substrate', branch = 'polkadot-v0.9.8', default-features = false }
 frame-system = { git = 'https://github.com/paritytech/substrate', branch = 'polkadot-v0.9.8', default-features = false }
-primitives = { path = '../../primitives/primitives', default-features = false }
+
+# orml
+orml-traits = { git = 'https://github.com/open-web3-stack/open-runtime-module-library', branch = 'master', default-features = false }
+
+# xcm
+xcm = { git = 'https://github.com/paritytech/polkadot', branch = 'release-v0.9.8', default-features = false }
 xcm-calls = { path = "../../primitives/xcm-calls", default-features = false }
+cumulus-pallet-xcm = { git = 'https://github.com/paritytech/cumulus', branch = 'polkadot-v0.9.8', default-features = false }
+
+# pint
+primitives = { path = '../../primitives/primitives', default-features = false }
 
 [features]
 default = ['std']
 std = [
+    "xcm/std",
     "frame-support/std",
     'frame-system/std',
+    "orml-traits/std",
     "primitives/std",
-    "xcm-calls/std"
+    "xcm-calls/std",
+    "cumulus-pallet-xcm/std"
 ]
